Rename: TableHeaderItem 컴포넌트 이동
