+++ src/components/Table/TableHeader/TableHeaderItem/index.ts
@@ -1,0 +1,1 @@
+export { TableHeaderItem } from './TableHeaderItem';
